refactor(ios): extract shared LayerShape from animated render view
